+++ chain/evm/evm_chain_test.go
@@ -1,0 +1,44 @@
+package evm_test
+
+import (
+	"testing"
+
+	chain_selectors "github.com/smartcontractkit/chain-selectors"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
+)
+
+func TestChain_ChainInfot(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name       string
+		selector   uint64
+		wantName   string
+		wantString string
+		wantFamily string
+	}{
+		{
+			name:       "returns correct info",
+			selector:   chain_selectors.ETHEREUM_MAINNET.Selector,
+			wantString: "ethereum-mainnet (5009297550715157269)",
+			wantName:   chain_selectors.ETHEREUM_MAINNET.Name,
+			wantFamily: chain_selectors.FamilyEVM,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			c := evm.Chain{
+				Selector: tt.selector,
+			}
+			assert.Equal(t, tt.selector, c.ChainSelector())
+			assert.Equal(t, tt.wantString, c.String())
+			assert.Equal(t, tt.wantName, c.Name())
+			assert.Equal(t, tt.wantFamily, c.Family())
+		})
+	}
+}