+++ deployment/changeset_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
+	"github.com/smartcontractkit/chainlink-deployments-framework/chain"
 	"github.com/smartcontractkit/chainlink-deployments-framework/datastore"
 )
 
@@ -95,7 +96,8 @@ func TestChangeSetLegacyFunction(t *testing.T) {
 
 func NewNoopEnvironment(t *testing.T) Environment {
 	t.Helper()
-	return *NewEnvironment(
+
+	return *NewCLDFEnvironment(
 		"noop",
 		logger.Test(t),
 		NewMemoryAddressBook(),
@@ -107,5 +109,6 @@ func NewNoopEnvironment(t *testing.T) Environment {
 		nil,
 		t.Context,
 		XXXGenerateTestOCRSecrets(),
+		chain.NewBlockChains(map[uint64]chain.BlockChain{}),
 	)
 }